+++ spec/requests/alchemy/site_requests_spec.rb
@@ -14,32 +14,5 @@ RSpec.describe 'Site requests' do
       get "http://#{site.host}/#{page.urlname}"
       expect(assigns(:current_alchemy_site).host).to eq(site.host)
     end
-
-    context 'with site_id of another site' do
-      let(:another_site) { Alchemy::Site.default }
-
-      let(:page) do
-        Alchemy::Site.current = another_site
-        root = create(:alchemy_page, :language_root, language: another_site.languages.last)
-        create(:alchemy_page, :public, parent: root)
-      end
-
-      context 'in params' do
-        it 'loads another site by id' do
-          get "http://#{site.host}/#{page.urlname}?site_id=#{another_site.id}"
-          expect(assigns(:current_alchemy_site).id).to eq(another_site.id)
-        end
-      end
-
-      context 'with site_id of another site in session' do
-        it 'loads another site by id' do
-          allow_any_instance_of(Alchemy::PagesController).to receive(:session) do
-            {site_id: another_site.id}
-          end
-          get "http://#{site.host}/#{page.urlname}"
-          expect(assigns(:current_alchemy_site).id).to eq(another_site.id)
-        end
-      end
-    end
   end
 end